+++ frontend/src/components/ui/StationDetailsWindow/reviews/common/HeaderWithRating.tsx
